Add files via upload
This project is a Python-based time series data scraper that collects data from a specified API endpoint following the Time Series Data Standards. It is designed to fetch time series data from a target website, parse the data, and export it to a CSV file.

Commit Message: Initial commit of the time series data scraper

Description: This commit includes the initial version of the Python script for scraping time series data from the specified API endpoint. The script makes HTTP requests, parses JSON responses, organizes the data, and exports it to a CSV file. This is the starting point of the project.